compleeted the team compare functions
