+++ app/ip3-app/.gitignore
@@ -21,3 +21,7 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+#csv
+seasonFixtures2122.csv
+seasonFixtures2223.csv
