+++ app/ip3-app/src/Component/TeamCompare.js
@@ -2,7 +2,6 @@ import { React, useState, useEffect } from "react";
 import Api from '../Helpers/Api';
 import SearchBar from "./SearchBar";
 import { useDrop } from "react-dnd";
-import PlayerCard from "./PlayerCard";
 import { TeamResultsCard } from "./TeamResultsCard";
 import TeamCard from "./TeamCard";
 
@@ -20,6 +19,11 @@ function TeamCompare(){
     const [dropTeam1, setDropTeam1] = useState("");
     const [dropTeam2, setDropTeam2] = useState("");
 
+    const [team1CompareResults, setTeam1CompareResults] = useState([]);
+    const [team2CompareResults, setTeam2CompareResults] = useState([]);
+
+
+
     useEffect(() => {
         const fetchData = async () => {
             try {
@@ -29,7 +33,7 @@ function TeamCompare(){
                     changeSelectionVisibility(false);
                 }
                 setTeams(data);
-                console.log(data);
+                // console.log(data);
             } catch (error) {
                 console.error(error);
             }
@@ -48,7 +52,7 @@ function TeamCompare(){
         console.log("search button clicked");
         Api.get(`team/list/${team}`)
             .then(res => {
-                console.log(res.data);
+                // console.log(res.data);
                 setTeams(res.data);
                 changeSelectionVisibility(true);
             })
@@ -63,7 +67,9 @@ function TeamCompare(){
             setBasket([]);
             setBasket((basket) =>
                 !basket.includes(item.team) ? [...basket, item.team] : basket);
-            selectedTeams.push(item.team);
+            // selectedTeams.push(item.team);
+            selectedTeams[0] = item.team;
+            findLast5Fix(selectedTeams);
             setDropTeam1(item.team);
         },
         collect: (monitor) => ({
@@ -78,7 +84,9 @@ function TeamCompare(){
             setBasket2([]);
             setBasket2((basket2) =>
                 !basket2.includes(item.team) ? [...basket2, item.team] : basket2);
-            selectedTeams.push(item.team);
+            // selectedTeams.push(item.team);
+            selectedTeams[1] = item.team;
+            findLast5Fix(selectedTeams);
             setDropTeam2(item.team);
         },
         collect: (monitor) => ({
@@ -105,6 +113,47 @@ function TeamCompare(){
         }
     }
 
+    const findLast5Fix = (teams) => {
+        if(teams.length == 2){
+
+            const team1 = teams[0].teamName;
+            const team2 = teams[1].teamName;
+
+            const last5Team1 = teams[0].fixtures.filter(f =>((f.awayTeamName == team1 || f.homeTeamName == team1) && (f.awayTeamName == team2 || f.homeTeamName == team2) && (f.fullTimeResult != '?'))).slice(0, 5)
+            const resultsTeam1 = new Array();
+            for (let i = 0; i < last5Team1.length; i++) {
+                if (last5Team1[i].fullTimeResult == 'D') {
+                    resultsTeam1.push('D');
+                }
+                else {
+                    if ((last5Team1[i].fullTimeResult == 'H' && team1 == last5Team1[i].homeTeamName) || (last5Team1[i].fullTimeResult == 'A' && team1 == last5Team1[i].awayTeamName))
+                    resultsTeam1.push('W')
+                    else resultsTeam1.push('L')
+                }
+            }
+            console.log("results for Team 1: "+resultsTeam1);
+            setTeam1CompareResults(resultsTeam1);
+
+            const last5Team2 = teams[1].fixtures.filter(f =>((f.awayTeamName == team1 || f.homeTeamName == team1) && (f.awayTeamName == team2 || f.homeTeamName == team2) && (f.fullTimeResult != '?'))).slice(0, 5)
+            const resultsTeam2 = new Array();
+            for (let i = 0; i < last5Team2.length; i++) {
+                if (last5Team2[i].fullTimeResult == 'D') {
+                    resultsTeam2.push('D');
+                }
+                else {
+                    if ((last5Team2[i].fullTimeResult == 'H' && team2 == last5Team2[i].homeTeamName) || (last5Team2[i].fullTimeResult == 'A' && team2 == last5Team2[i].awayTeamName))
+                    resultsTeam2.push('W')
+                    else resultsTeam2.push('L')
+                }
+            }
+            console.log("results for Team 2: "+resultsTeam2);
+            setTeam2CompareResults(resultsTeam2); 
+
+        } else if (teams.length > 2){
+            console.log("Too many selected teams for last 5 comparsions");
+        }
+    }
+
     return (
         <div className="Player" onClick={e => setToFalse(e)}>
             <SearchBar keyword={team} onChange={setTeam} fun={searchTeam} />
@@ -118,16 +167,16 @@ function TeamCompare(){
             <div className="CompareSection">
                 <div className="CardSections">
                     <div className={basket == "" ? "CardSection1" : "CardSection1WithBasket"} ref={dropRef}>
-                        {basket.map(team => <TeamCard team={team} />)}
+                        {basket.map(team => <TeamCard team={team} last5H2H={team1CompareResults}/>)}
                         {isOver && !basket.includes(dropTeam1) ? <p>Drop Player Here</p> : <p style={{display: "none"}}></p>}
                         {!isOver && !basket.includes(dropTeam1) ? <p>Search and Drag Player Here</p> : <p style={{display: "none"}}></p>}
 
-
                     </div>
                     <div className={basket2 == "" ? "CardSection2" : "CardSection2WithBasket"} ref={dropRef2}>
-                        {basket2.map(team => <TeamCard team={team} />)}  
+                        {basket2.map(team => <TeamCard team={team}  last5H2H={team2CompareResults}/>)}  
                         {isOver2 && !basket2.includes(dropTeam2) ? <p>Drop Player Here</p> : <p style={{display: "none"}}></p>}
                         {!isOver2 && !basket2.includes(dropTeam2) ? <p>Search and Drag Player Here</p> : <p style={{display: "none"}}></p>}
+
 
                     </div>
                 </div>
