+++ fixturedata/seasonFixtures2223.csv
@@ -173,3 +173,9 @@ SC0,18/03/2023,15:00,Celtic,Hibernian,3,1,H,0,1,A,S McLean,35,3,9,1,6,7,15,0,1,3
 SC0,18/03/2023,15:00,Dundee United,St Mirren,1,1,D,1,0,H,C Napier,11,16,2,3,11,9,1,4,1,1,0,0,2.63,3.3,2.7,2.6,3.25,2.6,2.65,3.3,2.65,2.86,3.37,2.62,2.75,3.1,2.6,2.7,3.1,2.63,2.87,3.37,2.7,2.73,3.24,2.61,2.1,1.73,2.18,1.74,2.28,1.77,2.16,1.68,0,1.98,1.88,2.04,1.87,2.04,1.94,1.95,1.87,2.7,3.25,2.63,2.7,3.1,2.65,2.75,3.2,2.65,2.77,3.24,2.8,2.75,3.1,2.6,2.63,3.2,2.63,2.77,3.25,2.81,2.69,3.15,2.69,2.3,1.62,2.33,1.65,2.4,1.66,2.28,1.61,0,1.9,1.95,1.94,1.96,1.94,1.99,1.9,1.92
 SC0,18/03/2023,15:00,Kilmarnock,St Johnstone,1,1,D,1,0,H,C Graham,14,9,4,2,10,21,9,5,3,5,0,0,2.15,3.3,3.5,2.15,3.25,3.4,2.2,3.3,3.35,2.22,3.2,3.76,2.15,3.1,3.5,2.15,3.1,3.5,2.22,3.35,3.76,2.17,3.19,3.51,2.3,1.62,2.47,1.59,2.47,1.62,2.34,1.58,-0.25,1.85,2,1.88,2.02,1.89,2.04,1.84,1.97,2.15,3.3,3.6,2.15,3.25,3.4,2.2,3.25,3.45,2.21,3.18,3.82,2.15,3.1,3.5,2.1,3.2,3.5,2.21,3.3,3.91,2.15,3.19,3.58,2.3,1.62,2.4,1.62,2.42,1.62,2.32,1.59,-0.25,1.8,2.05,1.88,2.04,1.89,2.09,1.83,2
 SC0,18/03/2023,15:00,Livingston,Ross County,2,1,H,2,0,H,W Collum,13,11,7,1,8,17,8,6,3,1,0,0,2.1,3.2,3.75,2.15,3.3,3.3,2.15,3.3,3.45,2.2,3.37,3.62,2.15,3.1,3.5,2.15,3.1,3.5,2.2,3.4,3.75,2.16,3.25,3.46,2.2,1.67,2.28,1.68,2.34,1.74,2.24,1.63,-0.25,1.85,2,1.88,2.02,1.91,2.04,1.85,1.97,2.1,3.1,3.75,2.05,3.2,3.6,2.15,3.25,3.5,2.13,3.27,3.94,2.15,3.1,3.5,2.1,3.25,3.4,2.2,3.32,3.94,2.13,3.19,3.59,2.3,1.62,2.33,1.65,2.35,1.7,2.28,1.61,-0.25,1.77,2.1,1.81,2.11,1.89,2.13,1.82,2.01
+SC0,01/04/2023,15:00,Hibernian,Motherwell,1,3,A,0,1,A,G Grainger,11,17,3,5,12,7,6,5,3,1,0,0,1.8,3.6,4.5,1.82,3.5,4.2,1.83,3.55,4.3,1.83,3.67,4.7,1.8,3.5,4.4,1.8,3.5,4.5,1.83,3.7,4.7,1.8,3.57,4.38,1.98,1.88,2.01,1.88,2.02,1.94,1.95,1.83,-0.5,1.77,2.02,1.83,2.09,1.83,2.09,1.79,2.01,1.85,3.5,4.5,1.83,3.5,4.2,1.83,3.55,4.3,1.86,3.57,4.69,1.83,3.5,4.2,1.85,3.5,4.33,1.88,3.63,4.69,1.84,3.49,4.32,2.1,1.73,2.13,1.78,2.17,1.82,2.04,1.76,-0.5,1.83,2.02,1.86,2.05,1.89,2.06,1.84,1.97
+SC0,01/04/2023,15:00,Kilmarnock,Hearts,2,1,H,2,1,H,E Anderson,17,10,4,3,12,15,2,5,2,2,1,0,2.9,3.5,2.38,2.85,3.4,2.3,3.05,3.5,2.25,3.05,3.51,2.41,2.9,3.4,2.3,2.9,3.4,2.38,3.05,3.55,2.41,2.93,3.44,2.33,1.9,1.95,1.93,1.95,1.93,1.99,1.86,1.92,0.25,1.8,2.05,1.84,2.08,1.85,2.09,1.8,2.01,2.6,3.4,2.7,2.55,3.3,2.6,3.05,3.5,2.25,2.66,3.52,2.72,2.6,3.25,2.62,2.6,3.3,2.7,3.05,3.52,2.72,2.61,3.36,2.64,1.93,1.93,1.93,1.96,2.04,1.97,1.92,1.87,0,1.9,1.95,1.93,1.98,2.15,2,1.9,1.91
+SC0,01/04/2023,15:00,Rangers,Dundee United,2,0,H,1,0,H,D Robertson,28,9,7,4,11,8,10,2,1,1,0,0,1.13,8,19,1.12,9.25,18,1.12,8.75,20,1.13,9.14,23.46,1.11,8,23,1.1,9.5,21,1.14,9.5,26,1.12,8.63,20.5,1.4,3,1.42,2.97,1.42,3.13,1.38,2.92,-2.5,2.05,1.75,2.11,1.78,2.12,1.83,2.03,1.77,1.11,11,15,1.11,9.5,20,1.12,8.75,20,1.11,10.4,28,1.1,8.5,23,1.1,9.5,21,1.12,11,29,1.1,9.46,22.62,1.33,3.4,1.37,3.2,1.4,3.4,1.34,3.15,-2.5,1.95,1.9,1.99,1.9,2.02,1.97,1.94,1.86
+SC0,01/04/2023,15:00,St Johnstone,Aberdeen,0,1,A,0,1,A,C Steven,14,12,6,3,6,8,7,7,3,2,1,1,2.9,3.2,2.45,2.87,3.2,2.4,2.9,3.25,2.5,3.22,3.21,2.46,3,3.1,2.4,3,3.1,2.45,3.22,3.3,2.5,3.01,3.17,2.42,2.2,1.67,2.27,1.68,2.27,1.72,2.19,1.66,0.25,1.77,2.1,1.83,2.09,1.83,2.17,1.76,2.06,3.2,3.25,2.25,3.2,3.2,2.25,2.9,3.25,2.5,3.53,3.19,2.32,3.25,3.1,2.25,3.3,3.2,2.25,3.53,3.3,2.5,3.27,3.17,2.28,2.3,1.62,2.35,1.65,2.35,1.69,2.26,1.62,0.25,1.93,1.93,1.93,1.96,1.97,2.05,1.88,1.93
+SC0,01/04/2023,15:00,St Mirren,Livingston,3,0,H,3,0,H,J Beaton,15,13,7,3,11,11,4,9,1,3,0,0,2.05,3.3,3.75,2.05,3.2,3.6,2.1,3.3,3.6,2.13,3.36,3.76,2.1,3.1,3.7,2.1,3.2,3.7,2.13,3.4,3.8,2.08,3.26,3.65,2.3,1.62,2.4,1.61,2.4,1.62,2.31,1.59,-0.25,1.75,2.05,1.83,2.09,1.84,2.13,1.78,2.04,2.05,3.2,3.8,2.05,3.1,3.75,2.1,3.3,3.6,2.16,3.11,4.12,2.05,3,4,2.05,3.2,3.9,2.16,3.4,4.12,2.1,3.1,3.83,2.63,1.5,2.69,1.51,2.69,1.6,2.54,1.5,-0.25,1.75,2.05,1.82,2.11,1.82,2.14,1.77,2.05
+SC0,02/04/2023,12:00,Ross County,Celtic,0,2,A,0,1,A,W Collum,8,28,2,9,7,6,2,9,0,0,0,0,17,7,1.14,14.5,8,1.16,14,7.5,1.17,16.51,8.71,1.16,17,7,1.15,15,8,1.13,19,9,1.17,15.42,7.87,1.15,1.4,3,1.39,3.07,1.41,3.1,1.39,2.9,2,2.05,1.75,2.12,1.78,2.15,1.81,2.08,1.75,17,8,1.14,14.5,8,1.16,14,7.5,1.17,18.65,9.21,1.15,17,7,1.15,15,8,1.15,19.9,9.27,1.17,16.24,8.04,1.15,1.36,3.2,1.38,3.14,1.4,3.41,1.37,3.03,2.25,1.85,2,1.89,2,1.93,2.05,1.86,1.96
